options in python and ruby plugins for cmd vs ext based package list extraction
