+++ crawler/crawler.conf
@@ -19,4 +19,4 @@
     [[ ruby_pkg ]]		
     		
     [[ python_pkg ]]		
-    avoid_setns = True
+    avoid_setns = False
